requirements.txt
slack_bolt==1.14.3
jira==3.1.1
slack-sdk
python-dotenv